TID is now long not int

+++ src/main/scala-2.11/ch/unibas/dmi/dbis/adam/index/IndexGenerator.scala
@@ -21,7 +21,7 @@ trait IndexGenerator {
    *
    */
   def index(indexname : IndexName, tablename : TableName, data: DataFrame): Index = {
-    val rdd = data.map { x => IndexerTuple(x.getInt(0), x.getSeq[VectorBase](1) : WorkingVector) }
+    val rdd = data.map { x => IndexerTuple(x.getLong(0), x.getSeq[VectorBase](1) : WorkingVector) }
     index(indexname, tablename, rdd)
   }
